+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-*
+* Fix HTML rendering of plain-text-only emails in the contents iframe
 
 ## [0.9.0] - 2026-04-12